Split import visibility integration tests
